fix data type of users_chat_id

+++ src/main/java/pro/sky/animal_shelter_ji22_team1_app/service/UserService.java
@@ -15,5 +15,5 @@ public interface UserService {
 
     void delete(Long userId);
 
-    UserEntity findByChatId(Integer chatId);
+    UserEntity findByChatId(Long chatId);
 }
